chore: restructure directories
