+++ deployer/pkg/deployer/app/packager/scanApp.go
